Downgrade undeclared extension log message to INFO
ECatcher seems to trigger on warnings, even if no exception is thrown (?).  This warning doesn't really need to be a warning if it's WAI and not something we're actually going to bother to reach out to registrars about (in this case, 101domain apparently isn't declaring the launch extension, but we let them use it anyway).

If at some point we decide to be stricter about declaring extensions and want to go determine how often this is happening, making it INFO vs WARNING doesn't make that any harder.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=130016606