Clean up LoadAndResaveCommand
This refactors LoadAndResaveCommand a bit so that it's more straightfoward
and just switches on an enum to determine the resource type.  This has the
pro of actually making the command line flag friendlier anyway.

Also changed how it handles a non-existent (or non-active) resource so that
it throws an IAE rather than silently doing stageEntityChange(null, null).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=134113671