Update README.md "Known issues" section
This seems like a better description of the actual contents of the section.  Expanded contents to mention the need for fully automated reporting and for crypto key storage.  Also reshuffled the list a bit to put the things likely to be top concerns first.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=136471298