Update the ordering of the DomainCreateFlow 'After Validation' extension point
Custom logic to block domain labels is currently being implemented in 'Before Save'. At that point, the default flow has already validated the SMD, premium, and reserved. Only then can we determine if we should block the label, so the 'After Validation' extension point doesn't currently fit the need.

However, as a result, if the label is blocked, but the fee extension is missing, a generic fee error is thrown, stating the fee extension must be present. We would rather state that the label is blocked. Then, if the SMD is present, we override the block, validate that the fee is present with the correct price, and allow the domain create.

The solution is to move the 'After Validation' extension point to actually be after all validation, including of the SMD, but before the fee check.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=143790163