Add build rules for "nocron" deployments
Add build rules for creating "nocron" war and ear files for use during
database restores in production and alpha.

After a little consideration, I think this is the right way to do this.  If we
want to set it up in other environments, we'll need to add these rules for
them, too, and at that point it may be worthwhile to wrap the rules we want in
a macro, but for now this is fairly clear and not execessively redundant.

This change works by modifying the zip_file rule to allow us to exclude
"cron.xml".  I have verified that:

- The alpha "nocron" files are generated without cron.xml.
- The production "nocron" files are generated without cron.xml.
- I can do a full push to alpha based on deploy_nocron which doesn't install a
  new cron.
- The normal production and alpha war-files _do_ contain cron.xml.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=153853205